+++ app/Http/Controllers/Pemesan/NegaraController.php
@@ -9,6 +9,7 @@ use App\Model\Pemesan\Kabupaten;
 use App\Model\Pemesan\Negara;
 use App\Model\Pemesan\Provinsi;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 
 use function Ramsey\Uuid\v1;
@@ -22,16 +23,20 @@ class NegaraController extends Controller
      */
     public function index()
     {
-        // SELECT n.id, n.nama,n.kode,p.nama as nm_prov, p.alias, k.prov_id,k.nama as nm_kab FROM negaras n JOIN provinsis p on n.id=p.negara_id JOIN kabupatens k on p.id=k.prov_id
-        $alamat = DB::table('negaras')
-            ->join('provinsis', 'negaras.id', '=', 'provinsis.negara_id')
-            ->join('kabupatens', 'provinsis.id', '=', 'kabupatens.prov_id')
-            ->select('kabupatens.nama as nm_kab', 'provinsis.nama as nm_prov', 'negaras.nama', 'negaras.kode')
-            ->paginate(10);
-        $n = Negara::paginate(5);
-        $p = Provinsi::paginate(5);
-        $k = Kabupaten::paginate(5);
-        return \view('pemesan.alamat.index', \compact('alamat', 'n', 'p', 'k'));
+        if (Auth::guard('pemesan')->check()) {
+            // SELECT n.id, n.nama,n.kode,p.nama as nm_prov, p.alias, k.prov_id,k.nama as nm_kab FROM negaras n JOIN provinsis p on n.id=p.negara_id JOIN kabupatens k on p.id=k.prov_id
+            $alamat = DB::table('negaras')
+                ->join('provinsis', 'negaras.id', '=', 'provinsis.negara_id')
+                ->join('kabupatens', 'provinsis.id', '=', 'kabupatens.prov_id')
+                ->select('kabupatens.nama as nm_kab', 'provinsis.nama as nm_prov', 'negaras.nama', 'negaras.kode')
+                ->paginate(10);
+            $n = Negara::paginate(5);
+            $p = Provinsi::paginate(5);
+            $k = Kabupaten::paginate(5);
+            return \view('pemesan.alamat.index', \compact('alamat', 'n', 'p', 'k'));
+        } else {
+            return \redirect()->route('login.index')->with(['msg' => 'anda harus login!!']);
+        }
     }
 
     /**
@@ -42,22 +47,15 @@ class NegaraController extends Controller
      */
     public function store(NegaraRequest $req)
     {
-        Negara::create([
-            'nama' => \ucwords($req->nama),
-            'kode' => $req->kode,
-        ]);
-        return \redirect()->back()->with(['msg' => "Berhasil menambah negara $req->nama"]);
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
+        if (Auth::guard('pemesan')->check()) {
+            Negara::create([
+                'nama' => \ucwords($req->nama),
+                'kode' => $req->kode,
+            ]);
+            return \redirect()->back()->with(['msg' => "Berhasil menambah negara $req->nama"]);
+        } else {
+            return \redirect()->route('login.index')->with(['msg' => 'anda harus login!!']);
+        }
     }
 
     /**
@@ -68,8 +66,12 @@ class NegaraController extends Controller
      */
     public function edit($id)
     {
-        $negara = Negara::find($id);
-        return \view('pemesan.alamat.negara.edit', \compact('negara'));
+        if (Auth::guard('pemesan')->check()) {
+            $negara = Negara::find($id);
+            return \view('pemesan.alamat.negara.edit', \compact('negara'));
+        } else {
+            return \redirect()->route('login.index')->with(['msg' => 'anda harus login!!']);
+        }
     }
 
     /**
@@ -81,11 +83,15 @@ class NegaraController extends Controller
      */
     public function update(NegaraReqUpdate $req,  $id)
     {
-        $negara = Negara::find($id);
-        $negara->nama = \ucwords($req->nama);
-        $negara->kode = $req->kode;
-        $negara->save();
-        return \redirect()->route('negara.index')->with(['msg' => "Berhasil mengubah negara $req->nama"]);
+        if (Auth::guard('pemesan')->check()) {
+            $negara = Negara::find($id);
+            $negara->nama = \ucwords($req->nama);
+            $negara->kode = $req->kode;
+            $negara->save();
+            return \redirect()->route('negara.index')->with(['msg' => "Berhasil mengubah negara $req->nama"]);
+        } else {
+            return \redirect()->route('login.index')->with(['msg' => 'anda harus login!!']);
+        }
     }
 
     /**
@@ -96,8 +102,12 @@ class NegaraController extends Controller
      */
     public function destroy($id)
     {
-        $negara = Negara::find($id);
-        $negara->delete();
-        return \redirect()->back()->with(['msg' => "Berhasil menghapus negara $negara->nama"]);
+        if (Auth::guard('pemesan')->check()) {
+            $negara = Negara::find($id);
+            $negara->delete();
+            return \redirect()->back()->with(['msg' => "Berhasil menghapus negara $negara->nama"]);
+        } else {
+            return \redirect()->route('login.index')->with(['msg' => 'anda harus login!!']);
+        }
     }
 }
